requirements.txt
geopandas
shapely
rtree
requests
rapidfuzz
tqdm
